rename mswindows to windows
